+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- **BC Break!**: Added `$page` and `$aliasNumberPages` parameters to
+  the `header()` and `footer()` functions.
 - Removed the `self::fail()` functions to unit tests.
 - Added the `getAliasNumberPages()` function to the `PdfDocument` class.
 - Minor modifications.